Backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
openmeteo_requests==1.3.0
openmeteo_sdk==1.19.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
rich==14.0.0
rich-toolkit==0.14.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sklearn-preprocessing==0.1.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
url-normalize==1.4.3
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
xgboost==3.0.2
